+++ packages/iov-core/src/index.ts
@@ -12,6 +12,18 @@ export {
   WalletSerializationString,
 } from "@iov/keycontrol";
 
-export { JsonRpcSigningServer } from "./jsonrpcsigningserver";
+export {
+  JsRpcClient,
+  jsRpcCode,
+  JsRpcErrorResponse,
+  JsRpcRequest,
+  JsRpcResponse,
+  JsRpcSuccessResponse,
+  parseJsRpcErrorResponse,
+  parseJsRpcId,
+  parseJsRpcRequest,
+  parseJsRpcResponse,
+} from "./jsrpc";
+export { JsRpcSigningServer } from "./jsrpcsigningserver";
 export { MultiChainSigner } from "./multichainsigner";
 export { GetIdentitiesAuthorization, SignAndPostAuthorization, SigningServerCore } from "./signingservercore";