+++ packages/iov-core/src/signingservice.spec.ts
@@ -8,6 +8,7 @@ import {
   Amount,
   ChainId,
   isConfirmedTransaction,
+  isPublicIdentity,
   PublicIdentity,
   PublicKeyBytes,
   SendTransaction,
@@ -17,15 +18,16 @@ import {
 import { bnsCodec, bnsConnector } from "@iov/bns";
 import { Ed25519, Random } from "@iov/crypto";
 import { Encoding } from "@iov/encoding";
-import {
-  JsonCompatibleDictionary,
-  JsonRpcClient,
-  JsonRpcResponse,
-  parseJsonRpcError,
-  parseJsonRpcResponse,
-  SimpleMessagingConnection,
-} from "@iov/jsonrpc";
+import { SimpleMessagingConnection } from "@iov/jsonrpc";
 import { firstEvent } from "@iov/stream";
+import {
+  JsRpcClient,
+  JsRpcCompatibleDictionary,
+  JsRpcRequest,
+  JsRpcResponse,
+  parseJsRpcErrorResponse,
+  parseJsRpcResponse,
+} from "./jsrpc";
 
 const { fromHex } = Encoding;
 
@@ -63,17 +65,19 @@ async function randomBnsAddress(): Promise<Address> {
   return bnsCodec.identityToAddress(randomIdentity);
 }
 
-function makeSimpleMessagingConnection(worker: Worker): SimpleMessagingConnection {
-  const producer: Producer<JsonRpcResponse> = {
+function makeSimpleMessagingConnection(
+  worker: Worker,
+): SimpleMessagingConnection<JsRpcRequest, JsRpcResponse> {
+  const producer: Producer<JsRpcResponse> = {
     start: listener => {
       // tslint:disable-next-line:no-object-mutation
       worker.onmessage = event => {
         // console.log("Got message from connection", event);
-        const responseError = parseJsonRpcError(event.data);
+        const responseError = parseJsRpcErrorResponse(event.data);
         if (responseError) {
           listener.next(responseError);
         } else {
-          const response = parseJsonRpcResponse(event.data);
+          const response = parseJsRpcResponse(event.data);
           listener.next(response);
         }
       };
@@ -115,6 +119,7 @@ describe("signingservice.worker", () => {
 
   it("can get bnsd identities", async () => {
     pendingWithoutBnsd();
+    pendingWithoutEthereum();
     pendingWithoutWorker();
 
     const bnsConnection = await bnsConnector(bnsdUrl).client();
@@ -122,9 +127,8 @@ describe("signingservice.worker", () => {
     const worker = new Worker(signingserviceKarmaUrl);
     await sleep(signingserviceBootTime);
 
-    const client = new JsonRpcClient(makeSimpleMessagingConnection(worker));
+    const client = new JsRpcClient(makeSimpleMessagingConnection(worker));
     const response = await client.run({
-      jsonrpc: "2.0",
       id: 123,
       method: "getIdentities",
       params: {
@@ -132,7 +136,6 @@ describe("signingservice.worker", () => {
         chainIds: [bnsConnection.chainId()],
       },
     });
-    expect(response.jsonrpc).toEqual("2.0");
     expect(response.id).toEqual(123);
     expect(response.result).toEqual(jasmine.any(Array));
     expect((response.result as ReadonlyArray<any>).length).toEqual(1);
@@ -154,9 +157,8 @@ describe("signingservice.worker", () => {
     const worker = new Worker(signingserviceKarmaUrl);
     await sleep(signingserviceBootTime);
 
-    const client = new JsonRpcClient(makeSimpleMessagingConnection(worker));
+    const client = new JsRpcClient(makeSimpleMessagingConnection(worker));
     const response = await client.run({
-      jsonrpc: "2.0",
       id: 123,
       method: "getIdentities",
       params: {
@@ -164,7 +166,6 @@ describe("signingservice.worker", () => {
         chainIds: [ganacheChainId],
       },
     });
-    expect(response.jsonrpc).toEqual("2.0");
     expect(response.id).toEqual(123);
     expect(response.result).toEqual(jasmine.any(Array));
     expect((response.result as ReadonlyArray<any>).length).toEqual(1);
@@ -174,6 +175,7 @@ describe("signingservice.worker", () => {
   });
 
   it("can get BNS or Ethereum identities", async () => {
+    pendingWithoutBnsd();
     pendingWithoutEthereum();
     pendingWithoutWorker();
 
@@ -182,9 +184,8 @@ describe("signingservice.worker", () => {
     const worker = new Worker(signingserviceKarmaUrl);
     await sleep(signingserviceBootTime);
 
-    const client = new JsonRpcClient(makeSimpleMessagingConnection(worker));
+    const client = new JsRpcClient(makeSimpleMessagingConnection(worker));
     const response = await client.run({
-      jsonrpc: "2.0",
       id: 123,
       method: "getIdentities",
       params: {
@@ -192,7 +193,6 @@ describe("signingservice.worker", () => {
         chainIds: [ganacheChainId, bnsConnection.chainId()],
       },
     });
-    expect(response.jsonrpc).toEqual("2.0");
     expect(response.id).toEqual(123);
     expect(response.result).toEqual(jasmine.any(Array));
     expect((response.result as ReadonlyArray<any>).length).toEqual(2);
@@ -209,6 +209,7 @@ describe("signingservice.worker", () => {
 
   it("send a signing request to service", async () => {
     pendingWithoutBnsd();
+    pendingWithoutEthereum();
     pendingWithoutWorker();
 
     const bnsConnection = await bnsConnector(bnsdUrl).client();
@@ -216,10 +217,9 @@ describe("signingservice.worker", () => {
     const worker = new Worker(signingserviceKarmaUrl);
     await sleep(signingserviceBootTime);
 
-    const client = new JsonRpcClient(makeSimpleMessagingConnection(worker));
+    const client = new JsRpcClient(makeSimpleMessagingConnection(worker));
 
     const identitiesResponse = await client.run({
-      jsonrpc: "2.0",
       id: 1,
       method: "getIdentities",
       params: {
@@ -227,7 +227,12 @@ describe("signingservice.worker", () => {
         chainIds: [bnsConnection.chainId()],
       },
     });
-    const signer: PublicIdentity = identitiesResponse.result[0];
+    expect(identitiesResponse.result).toEqual(jasmine.any(Array));
+    expect((identitiesResponse.result as ReadonlyArray<any>).length).toEqual(1);
+    const signer = identitiesResponse.result[0];
+    if (!isPublicIdentity(signer)) {
+      throw new Error("Identity element is not valid");
+    }
 
     const send: SendTransaction = {
       kind: "bcp/send",
@@ -238,12 +243,13 @@ describe("signingservice.worker", () => {
     };
 
     const signAndPostResponse = await client.run({
-      jsonrpc: "2.0",
       id: 2,
       method: "signAndPost",
       params: {
         reason: "Please sign",
-        transaction: (send as unknown) as JsonCompatibleDictionary,
+        // Cast needed since type of indices of transaction is not string at compile time.
+        // see https://stackoverflow.com/a/37006179/2013738
+        transaction: (send as unknown) as JsRpcCompatibleDictionary,
       },
     });
     const transactionId: TransactionId = signAndPostResponse.result;