+++ packages/iov-jsonrpc/types/jsonrpcclient.d.ts
@@ -1,11 +1,11 @@
 import { Stream } from "xstream";
 import { JsonRpcRequest, JsonRpcResponse, JsonRpcSuccessResponse } from "./types";
-export interface SimpleMessagingConnection {
-    readonly responseStream: Stream<JsonRpcResponse>;
-    readonly sendRequest: (request: JsonRpcRequest) => void;
+export interface SimpleMessagingConnection<Request, Response> {
+    readonly responseStream: Stream<Response>;
+    readonly sendRequest: (request: Request) => void;
 }
 export declare class JsonRpcClient {
     private readonly connection;
-    constructor(connection: SimpleMessagingConnection);
+    constructor(connection: SimpleMessagingConnection<JsonRpcRequest, JsonRpcResponse>);
     run(request: JsonRpcRequest): Promise<JsonRpcSuccessResponse>;
 }